feat(06-03): DEBT-04 propagate origin_dtype through _set_rgb per-channel branch (Task 5)
Closes DEBT-04 (CONTEXT.md D-11): before this commit, setting an
individual RGB channel (e.g. ``pcd.scalar_fields['red'] = red_field``)
silently re-derived a new ``DtypeState`` from the assigned uint8 array,
losing the source-dtype contract used by ``revert_sf_types=True``
PLY/LAS writers downstream.

Changes:

* ``ScalarFieldManager._set_rgb`` per-channel branch (scalars) now
  passes ``origin_dtype`` to :meth:`RGBFields.initialize` when
  auto-creating the RGB triplet on first-channel-set. Mirrors the
  sibling ``_set_normals`` shape and aligns with the rest of
  ``__setitem__``'s origin_dtype-aware code paths.
* :meth:`ScalarFieldTriplet.initialize` grows an ``origin_dtype:
  SfOrigDtT = None`` parameter (default-None preserves all existing
  call sites — ``RGBFields.initialize(N)`` / ``NormalFields.initialize(N)``).
* The ``# TODO update initialize to receive and origin_dtype ...``
  comment at ``scalar_field_manager.py:643`` is deleted.

Regression test (``tests/scalar_fields/test_scalar_field_manager.py``):

* ``test_rgb_origin_dtype_set_get_per_channel`` — in-memory check that
  origin_dtype is preserved after per-channel set.
* ``test_rgb_origin_dtype_set_get_per_channel_roundtrip`` — full
  save→reload integration through PLY with ``revert_sf_types=True``,
  asserting the on-disk dtype matches the source ``origin_dtype``.

Refs: DEBT-04, CONTEXT.md D-11

+++ src/pchandler/scalar_fields/scalar_field_manager.py
@@ -640,10 +640,12 @@ class ScalarFieldManager:
         if name in RGB_NAMES.float:
             value = cast(Array_Uint8_T, normalize_uint8(value))
 
-        # TODO update initialize to receive and origin_dtype in case it's being defined by a sequence of vectors
         if name in RGB_NAMES.scalars:
             if self.rgb is None:
-                self.rgb = RGBFields.initialize(self.num_points)
+                # DEBT-04: propagate origin_dtype on first-channel-set so the
+                # field's source dtype is preserved across save+reload (mirrors
+                # the sibling :meth:`_set_normals` per-channel branch below).
+                self.rgb = RGBFields.initialize(self.num_points, origin_dtype=origin_dtype)
 
             index = RGB_NAMES.get_position(name)
             self.rgb.arr[:, index] = Vector_Uint8_T(value)  # Perform validation as it's being assigned directly
